requirements.txt
cbpro
matplotlib
pandas
Pillow
board